requirements.txt
WTForms
pandas
numpy
Flask
matplotlib
Werkzeug~=2.0.0
Jinja2
scikit-learn
seaborn
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy
sqlalchemy
email_validator
python-decouple
gunicorn
docx-mailmerge
docxcompose
docx
category_encoders
scipy
nltk
mysql
mysql-connector-python
skforecast
statsmodels~=0.13.2
plotly~=5.7.0
requests
